Add quick fix to integrator attribute trouble
This is a very quick fix that is not permanent and absolutely not
intended to be used in production. It simply sets the appropriate
attributes to the values the should have if the nonequilibrium
integrator is unmetropolized Langevin and not measuring shadow work or
heat (something that should be common). As soon as a fix in openmmtools
is finalized, this can be removed.