Write class to automate construction of both legs of FEP
This is the first commit for the class that will automate both legs of
FEP. Specifically, whenever we do a complex relative calculation, we
will also want to do a solvent version, with the same atom map. This
class will facilitate that, and in this commit, now gains the ability to
construct the complex phase.