Add tasklet for writing nonequilibrium results; begin type hinting
In this commit, I am adding a tasklet for writing nonequilibrium results
to disk. This allows us to asynchronously write files (and relieve the
code from being in the driver class, allowing celery to schedule it). A
similar tasklet will be created for equilibrium results. Note that for
now, we are not writing anything fancy; in the future, this will be
replaced by a more clever storage backend.

The other modification in this commit is the beginning of the addition
of type hints. These are useful for IDEs and type checkers such as mypy
for catching bugs related to types at develop-time. As the use of these
type hints expands in the project, we will add mypy checking to the
travis tests.