Begin refactor of relative_setup
Now that the writing is being deferred to tasklets, this commit begins
by removing the trajectory writing, etc. It also provides some cleanup
by using dictionaries and iterators instead of duplicating the same
code, which will help avoid copy/paste errors. There are still several
remaining tasks:

- Complete the refactor by changing the code in the run_iteration method
- Add another tasklet to compute equivalent reduced potential for
  nonalchemical endpoints.

The second is necessary because we are no longer returning the results
of the simulations to the driver, beyond the floats representing reduced
potentials and work values. This enables the driver to run tasks
asynchronously, and is more tolerant of high-latency network
connections.